9P: change in _9p_rerror's prototype

+++ src/FSAL/FSAL_VFS/fsal_xattrs.c
@@ -876,7 +876,7 @@ fsal_status_t VFSFSAL_SetXAttrValue(fsal_handle_t * p_objecthandle,  /* IN */
 
   len = strnlen((char *)buffer_addr, buffer_size);
 
-  printf( "VFSFSAL_SetXAttrValue size=%llu len=%llu\n", buffer_size, len ) ;
+  printf( "VFSFSAL_SetXAttrValue size=%llu len=%llu\n", (unsigned long long)buffer_size, (unsigned long long)len ) ;
 
   TakeTokenFSCall();
   if(len == 0)
